Create rotable module
